requirements.txt
alabaster==0.7.12
altair==4.1.0
appdirs==1.4.4
argh==0.26.2
astor==0.8.1
astroid==2.3.3
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.4.3
Babel==2.8.0
backcall==0.2.0
base58==2.0.1
bcrypt==3.1.7
bleach==3.1.5
blinker==1.4
boto3==1.14.7
botocore==1.17.7
brotlipy==0.7.0
cachetools==4.1.0
category-encoders==2.2.2
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
docutils==0.16
entrypoints==0.3
enum-compat==0.0.3
filelock==3.0.12
flake8==3.7.9
future==0.18.2
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.1
intervaltree==3.0.2
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.4
jupyterlab-server==1.1.5
keyring==21.1.1
kiwisolver==1.2.0
lazy-object-proxy==1.5.0
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.0
numpydoc==1.0.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pathtools==0.1.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pipenv==2020.6.2
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
pycodestyle==2.5.0
pycparser==2.20
pydeck==0.4.0b2
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.4.4
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.4
QtPy==1.9.0
requests==2.24.0
rope==0.17.0
Rtree==0.9.4
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.5.0
Send2Trash==1.5.0
six==1.15.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
Sphinx==3.0.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.3
spyder-kernels==1.9.1
statsmodels==0.11.1
streamlit==0.61.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.1
ujson==1.35
urllib3==1.25.9
validators==0.15.0
virtualenv==20.0.25
virtualenv-clone==0.5.4
watchdog==0.10.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
yapf==0.28.0
zipp==3.1.0
